+++ src/store/action-creators/index.ts
@@ -1,13 +1,13 @@
 import * as BoardActionCreators from './board';
 import * as ColumnActionCreators from './column';
-import * as BordListActionCreators from './main';
+import * as GetBordActionCreators from './getBoards';
 import * as DeleteBordActionCreators from './deleteBoard';
 import * as SetBordActionCreators from './setBoard';
 
 export default {
   ...BoardActionCreators,
   ...ColumnActionCreators,
-  ...BordListActionCreators,
+  ...GetBordActionCreators,
   ...DeleteBordActionCreators,
   ...SetBordActionCreators,
 };
